feat : 전체 로또 티켓 생성 테스트 및 기능 추가, 관련 소스 수정

+++ src/main/java/lotto/domain/vending/TicketAmount.java
@@ -30,7 +30,7 @@ public class TicketAmount {
         return autoCount;
     }
 
-    public float total() {
+    public int total() {
         return manualCount + autoCount;
     }
 }
